new carousel layout isolated and added to the CMS
