+++ components/layouts/featured-topic.vue
@@ -8,105 +8,17 @@ const props = defineProps({
   },
   maxItems: {
     type: Number,
-    default: 15,
+    default: 4,
   },
 });
 
 const reactiveItems = toRef(props.list, "listItems");
 const { breakpoint } = useBreakpoints();
 const isLgBreakpoint = computed(() => breakpoint("<lg"));
-
-const getImgSizesBasedOnItemImgRatio = (item, obj) => {
-  const imgHeight = Number(item.image?.height);
-  const imgWidth = Number(item.image?.width);
-
-  // Calcluate Aspect Ratio (Width / Height)
-  let ratio = 1;
-  if (
-    imgHeight &&
-    imgWidth &&
-    !isNaN(imgHeight) &&
-    !isNaN(imgWidth) &&
-    imgHeight !== 0
-  ) {
-    ratio = imgWidth / imgHeight;
-  }
-
-  // Treat obj values as HEIGHT (user request), calculate WIDTH
-  const sizeObj = {
-    xs: obj.xs ? [Math.round(obj.xs * ratio), obj.xs] : undefined,
-    md: obj.md ? [Math.round(obj.md * ratio), obj.md] : undefined,
-    lg: obj.lg ? [Math.round(obj.lg * ratio), obj.lg] : undefined,
-  };
-
-  // Clean up undefined
-  Object.keys(sizeObj).forEach(
-    (key) => sizeObj[key] === undefined && delete sizeObj[key]
-  );
-
-  return sizeObj;
-};
 </script>
 
 <template>
   <div class="layout layout-featured-topic">
-    <div class="layout-tile mb-8">
-      <div class="flex flex-wrap gap-3">
-        <div
-          v-for="(item, index) in reactiveItems?.slice(0, props.maxItems)"
-          :key="`carousel-${item.id}-${index}`"
-          class="item"
-          style="height: -webkit-fill-available"
-        >
-          <MediaCard
-            inCarousel
-            showTease
-            isVertical
-            :data="item"
-            :allowVerticalEffect="false"
-            imgCol="h-12rem md:h-14rem lg:h-20rem"
-            :size="
-              getImgSizesBasedOnItemImgRatio(item, {
-                xs: 192,
-                md: 224,
-                lg: 320,
-              })
-            "
-            @on-click="dynamicNavigation(item)"
-          />
-        </div>
-      </div>
-    </div>
-    <MaterialCarouselBasic
-      :enableThrow="true"
-      :items-to-show="2"
-      :min-content-width="180"
-      :gap="16"
-    >
-      <div
-        v-for="(item, index) in reactiveItems?.slice(0, props.maxItems)"
-        :key="`carousel-${item.id}-${index}`"
-        class="item"
-        style="height: -webkit-fill-available"
-      >
-        <MediaCard
-          inCarousel
-          showTease
-          isVertical
-          :data="item"
-          :allowVerticalEffect="false"
-          imgCol="h-12rem md:h-14rem lg:h-20rem"
-          :size="
-            getImgSizesBasedOnItemImgRatio(item, {
-              xs: 192,
-              md: 224,
-              lg: 320,
-            })
-          "
-          @on-click="dynamicNavigation(item)"
-        />
-      </div>
-    </MaterialCarouselBasic>
     <div
       class="ad mb-5 col-12 flex align-items-center justify-content-center lg:hidden"
     >
